Refactor part of the backend and fix some additional bux
