+++ server/app/main.py
@@ -7,6 +7,7 @@ from app.routes.file import router as file_router
 from app.routes.graph import router as graph_router
 from app.routes.study import router as study_router
 from app.routes.history import router as history_router
+from app.routes.comment import router as comment_router
 
 import os
 
@@ -20,6 +21,7 @@ def get_application():
     app.include_router(graph_router, prefix="/graph")
     app.include_router(study_router, prefix="/study")
     app.include_router(history_router, prefix="/history")
+    app.include_router(comment_router, prefix="/history/comment")
 
     app.add_middleware(
         CORSMiddleware,
